examples/ema_expert_bybit/pyproject.toml
[project]
name = "ema-expert-bybit-examples"
version = "0.1.0"
description = "EMA expert for Bybit"
authors = [{ name = "Oleg Alexandrov", email = "alexandrovoleg.ru@gmail.com" }]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "algoexpert",
    "pandas",
    "python-dotenv",
]

[tool.uv.sources]
algoexpert = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"